+++ src/app/HISTORY.md
@@ -112,6 +112,10 @@ App Framework Change History
   functionality as the old behavior, pass your HTML string through
   `Y.Node.create()` before passing it to `container`.
 
+* [!] Destroying a view no longer also destroys the view's container node by
+  default. To destroy a view's container node when destroying the view, pass
+  `{remove: true}` to the view's `destroy()` method. [Ticket #2531689]
+
 * Added a View extension, `Y.View.NodeMap`, that can be mixed into a `View`
   subclass to provide a static `getByNode()` method that returns the nearest
   View instance associated with a given Node (similar to `Widget.getByNode()`).